--- pve/src/main/java/me/hsgamer/flexegames/game/pve/feature/ConfigFeature.java
@@ -1,7 +1,0 @@
-package me.hsgamer.flexegames.game.pve.feature;
-
-import me.hsgamer.flexegames.game.pve.PveGameConfig;
-import me.hsgamer.minigamecore.base.Feature;
-
-public record ConfigFeature(PveGameConfig config) implements Feature {
-}
