+++ pve/src/main/java/me/hsgamer/flexegames/game/pve/mob/ArenaMob.java
@@ -1,6 +1,7 @@
 package me.hsgamer.flexegames.game.pve.mob;
 
-import me.hsgamer.flexegames.game.pve.feature.ConfigFeature;
+import me.hsgamer.flexegames.feature.arena.GameFeature;
+import me.hsgamer.flexegames.game.pve.PveProperties;
 import me.hsgamer.flexegames.game.pve.feature.StageFeature;
 import me.hsgamer.minigamecore.base.Arena;
 import net.kyori.adventure.text.Component;
@@ -28,7 +29,8 @@ public class ArenaMob extends EntityCreature {
         super(entityType);
         this.arena = arena;
         var stage = getStage();
-        int multi = arena.getFeature(ConfigFeature.class).config().isToughMob() ? 2 : 1;
+        boolean toughMob = arena.getFeature(GameFeature.class).propertyMap().getProperty(PveProperties.TOUGH_MOB);
+        int multi = toughMob ? 2 : 1;
         getAttribute(Attribute.MAX_HEALTH).setBaseValue((getMaxHealth() + stage * 2) * multi);
         getAttribute(Attribute.ATTACK_DAMAGE).setBaseValue((1 + stage / 4f) * multi);
         heal();
